+++ src/__test-utils__/helpers.tsx
@@ -1,19 +1,20 @@
+import React from "react";
+
 import { ThemeProvider } from "@mui/material/styles";
 import { DiamondTheme } from "../themes/DiamondTheme";
 import { render, RenderResult } from "@testing-library/react";
 import { ThemeProviderProps } from "@mui/material/styles/ThemeProvider";
 
-export const renderWithProviders = (
-  ui: React.ReactNode,
-  themeOptions?: Omit<ThemeProviderProps, "theme">,
-): RenderResult => {
-  const Wrapper = ({ children }: { children: React.ReactNode }) => {
-    return (
-      <ThemeProvider {...themeOptions} theme={DiamondTheme}>
-        {children}
-      </ThemeProvider>
-    );
-  };
+type ThemeProviderPropsWithOptionalTheme = Omit<ThemeProviderProps, "theme"> &
+  Partial<Pick<ThemeProviderProps, "theme">>;
 
-  return render(ui, { wrapper: Wrapper });
+export const renderWithProviders = (
+  children: React.ReactNode,
+  themeOptions?: ThemeProviderPropsWithOptionalTheme,
+): RenderResult => {
+  return render(
+    <ThemeProvider theme={DiamondTheme} {...themeOptions}>
+      {children}
+    </ThemeProvider>,
+  );
 };