+++ src/components/controls/ImageColourSchemeSwitch.test.tsx
@@ -7,6 +7,7 @@ import { screen } from "@testing-library/react";
 describe("ImageColourSchemeSwitch", () => {
   const testVals = {
     src: "src/light",
+    srcDark: "src/dark",
     alt: "test-alt",
   };
 
@@ -59,14 +60,21 @@ describe("ImageColourSchemeSwitch", () => {
   });
 
   it("should have alternate src", () => {
-    const srcDark = "src/dark";
-
-    renderWithProviders(
-      <ImageColourSchemeSwitch image={{ ...testVals, srcDark }} />,
-      { defaultMode: "dark" },
-    );
+    renderWithProviders(<ImageColourSchemeSwitch image={{ ...testVals }} />, {
+      defaultMode: "dark",
+    });
     const img = screen.getByTestId("image-dark");
+    expect(img).toHaveAttribute("src", testVals.srcDark);
+  });
 
-    expect(img).toHaveAttribute("src", srcDark);
+  it("should have src when no srcDark set but dark mode selected", () => {
+    renderWithProviders(
+      <ImageColourSchemeSwitch image={{ ...testVals, srcDark: undefined }} />,
+      {
+        defaultMode: "dark",
+      },
+    );
+    const img = screen.getByRole("img");
+    expect(img).toHaveAttribute("src", testVals.src);
   });
 });