+++ changelog.md
@@ -6,19 +6,23 @@ SciReactUI Changelog
 --------------------
 
 ### Added
--
+- Logo component, to easily add the theme logo to anywhere
 
 ### Fixed
 - 
 
 ### Changed
-- Breaking changes: ImageColorSchemeSwitch, ImageColorSchemeSwitchType and ImageColorSchemeSwitchProps renamed to ImageColourSchemeSwitch, ImageColourSchemeSwitchType and ImageColourSchemeSwitchProps. User component color prop renamed to colour.
+- Breaking change: ImageColorSchemeSwitch, ImageColorSchemeSwitchType and ImageColorSchemeSwitchProps renamed to ImageColourSchemeSwitch, ImageColourSchemeSwitchType and ImageColourSchemeSwitchProps. 
+- Breaking change: User component color prop renamed to colour.
+- ImageColourSchemeSwitch takes a parameter to interchange images on colour scheme switch.
+
 
 [v0.2.0] - 2025-06-11
 ---------------------
 
 ### Fixed
 - Styles added to Navbar and Footer incorrectly remove built in styles.
+- Logo not appearing when no dark src set in dark mode.
 
 ### Changed
 - Breadcrumbs component takes optional linkComponent prop for page routing.